Add modern logging system
This adds the `simlog` type to handle most diagnostic output,
which replaces procedures from `common_io`. This is aggregated
into a new `mfe_env` type which is instantiated in the main
program and passed into the object hierarchy starting with
the `mfe_sim` object. This `mfe_env` type will hold additional
common simulation environment data in the future.