WIP: refactored to allow for pde plugin
This sets the stage to turn the PDE RHS method into a run time plugin.
It adds a `pde` abstract class that the code operates on, but
initialization hardwires it to be the existing Navier-Stokes equations.

It also uses a length-type parameter for `pde` that is the number of
equations. This works for Intel, but triggers an ICE for NAG. This
was speculative; the final implementation was unlikely to use this
approach anyway because PDT support in gfortran is lacking.