Parametrize the cell_data type
This adds npde as a parameter and makes all the component arrays
non-allocatable of the proper size.

NB: The cdata dummy arguments to the specific pde%rhs procedures
must be declared as type(cell_data(n)) where "n" is the actual
number.  If it is declared assumed-length (*) Intel segfaults (ICE).